+++ lib/td_df_lib/template_factory.ex
@@ -5,6 +5,33 @@ defmodule TdDfLib.TemplateFactory do
 
   defmacro __using__(_opts) do
     quote do
+      def template_factory(%{cardinality: cardinality} = attrs) do
+        %{
+          name: sequence("template_name"),
+          scope: "dd",
+          label: sequence("template_label"),
+          id: sequence(:template_id, &(&1 + 999_000)),
+          updated_at: DateTime.utc_now(),
+          content: [
+            %{
+              "name" => sequence("group_name"),
+              "fields" => [
+                build(:template_field,
+                  name: "string",
+                  cardinality: cardinality
+                ),
+                build(:template_field,
+                  name: "list",
+                  type: "list",
+                  values: %{"fixed" => ["one", "two", "three"]}
+                )
+              ]
+            }
+          ]
+        }
+        |> merge_attributes(attrs)
+      end
+
       def template_factory(attrs) do
         %{
           name: sequence("template_name"),